worker/requirements.txt
kafka-python==2.0.2
pymysql==1.1.0
nltk==3.8.1
python-dotenv==1.0.0
